+++ clang-tidy/ClangTidyForceLinker.h
@@ -50,6 +50,11 @@ extern volatile int CppCoreGuidelinesModuleAnchorSource;
 static int LLVM_ATTRIBUTE_UNUSED CppCoreGuidelinesModuleAnchorDestination =
     CppCoreGuidelinesModuleAnchorSource;
 
+// This anchor is used to force the linker to link the FPGAModule.
+extern volatile int FPGAModuleAnchorSource;
+static int LLVM_ATTRIBUTE_UNUSED FPGAModuleAnchorDestination =
+    FPGAModuleAnchorSource;
+
 // This anchor is used to force the linker to link the FuchsiaModule.
 extern volatile int FuchsiaModuleAnchorSource;
 static int LLVM_ATTRIBUTE_UNUSED FuchsiaModuleAnchorDestination =