+++ docs/ReleaseNotes.rst
@@ -67,6 +67,16 @@ The improvements are...
 Improvements to clang-tidy
 --------------------------
 
+- New :doc:`fpga <clang-tidy/index` module
+
+  Contains checks related to OpenCL programming for FPGAs.
+
+- New :doc:`fpga-struct-pack-align
+  <clang-tidy/checks/struct-pack-align>` check
+
+  Finds structs that are inefficiently packed or aligned and recommends
+  packing and/or aligning of said structs as needed.
+
 - New :doc:`bugprone-dynamic-static-initializers
   <clang-tidy/checks/bugprone-dynamic-static-initializers>` check.